refactor scanner into data pipeline
